+++ spec/board_spec.rb
@@ -1,5 +1,6 @@
-require 'spec_helper'
+require_relative 'spec_helper'
 require_relative '../lib/board'
+require_relative  '../lib/player'
 
 describe Board do
   let(:board) { Board.new }
